requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
pydantic
passlib[bcrypt]
python-jose
python-dotenv
pyjwt
pydantic_settings